split screen pacakge
