+++ core/src/tdt4240/tankathon/game/gamescreens/LoadingScreen.kt
@@ -1,4 +1,4 @@
-package tdt4240.tankathon.game.screens
+package tdt4240.tankathon.game.gamescreens
 
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.graphics.GL20
@@ -9,16 +9,15 @@ import com.badlogic.gdx.maps.objects.RectangleMapObject
 import com.badlogic.gdx.maps.tiled.TiledMap
 import com.badlogic.gdx.math.Rectangle
 import com.badlogic.gdx.math.Vector2
-import ktx.graphics.use
 import ktx.log.info
 import ktx.log.logger
-import tdt4240.tankathon.game.GameManager
 import tdt4240.tankathon.game.TankathonGame
-import tdt4240.tankathon.game.V_HEIGHT_PIXELS
+import tdt4240.tankathon.game.gamescreens.AbstractScreen
+import tdt4240.tankathon.game.gamescreens.GameScreen
 
 private val LOG = logger<LoadingScreen>()
 
-class LoadingScreen(game: TankathonGame) : AbstractUI(game){
+class LoadingScreen(game: TankathonGame) : AbstractScreen(game){
     override fun show(){
         menuStage.clear()
         engine.removeAllEntities()
@@ -43,10 +42,6 @@ class LoadingScreen(game: TankathonGame) : AbstractUI(game){
     override fun render(delta: Float) {
         Gdx.gl.glClearColor(0f, 0f, 0.2f, 1f)
         Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT)
-        batch.use {
-            val str = "Loading..."
-            uiFont.draw(it, str, 0f, 40f)
-        }
     }
 
     private fun parseCollision(tiledMap: TiledMap){
