+++ core/src/tdt4240/tankathon/game/menuscreens/AbstractUI.kt
@@ -1,21 +1,17 @@
-package tdt4240.tankathon.game.screens
+package tdt4240.tankathon.game.menuscreens
 
 import com.badlogic.gdx.Gdx
 import com.badlogic.gdx.graphics.GL20
 import com.badlogic.gdx.graphics.g2d.BitmapFont
 import com.badlogic.gdx.graphics.g2d.TextureAtlas
 import com.badlogic.gdx.math.Vector3
-import com.badlogic.gdx.scenes.scene2d.Actor
 import com.badlogic.gdx.scenes.scene2d.ui.Label
 import com.badlogic.gdx.scenes.scene2d.ui.Skin
 import com.badlogic.gdx.scenes.scene2d.ui.Table
-import com.badlogic.gdx.scenes.scene2d.ui.TextButton
-import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener
 import com.badlogic.gdx.utils.Align
 import ktx.graphics.use
-import tdt4240.tankathon.game.GameManager
 import tdt4240.tankathon.game.TankathonGame
-import java.lang.reflect.Type
+import tdt4240.tankathon.game.gamescreens.AbstractScreen
 
 abstract class AbstractUI(
         /**
