+++ core/src/tdt4240/tankathon/game/menuscreens/ScoreBoardScreen.kt
@@ -1,18 +1,9 @@
-package tdt4240.tankathon.game.screens
+package tdt4240.tankathon.game.menuscreens
 
 import com.badlogic.gdx.Gdx
-import com.badlogic.gdx.graphics.GL20
-import com.badlogic.gdx.graphics.OrthographicCamera
-import com.badlogic.gdx.graphics.g2d.BitmapFont
-import com.badlogic.gdx.graphics.g2d.TextureAtlas
-import com.badlogic.gdx.math.Vector3
 import com.badlogic.gdx.scenes.scene2d.Actor
 import com.badlogic.gdx.scenes.scene2d.ui.*
 import com.badlogic.gdx.scenes.scene2d.utils.ChangeListener
-import com.badlogic.gdx.utils.Align
-import com.badlogic.gdx.utils.viewport.FitViewport
-import com.badlogic.gdx.utils.viewport.Viewport
-import ktx.graphics.use
 import ktx.log.info
 import ktx.log.logger
 import tdt4240.tankathon.game.*
